System: rename UserFieldGateway to CustomFieldGateway, update namespaces
