+++ modules/User Admin/userFields.php
@@ -19,7 +19,7 @@ along with this program. If not, see <http://www.gnu.org/licenses/>.
 
 use Gibbon\Tables\DataTable;
 use Gibbon\Services\Format;
-use Gibbon\Domain\User\UserFieldGateway;
+use Gibbon\Domain\System\CustomFieldGateway;
 
 if (isActionAccessible($guid, $connection2, '/modules/User Admin/userFields.php') == false) {
     // Access denied
@@ -28,7 +28,7 @@ if (isActionAccessible($guid, $connection2, '/modules/User Admin/userFields.php'
     //Proceed!
     $page->breadcrumbs->add(__('Manage Custom Fields'));
 
-    $userFieldGateway = $container->get(UserFieldGateway::class);
+    $userFieldGateway = $container->get(CustomFieldGateway::class);
     
     // QUERY
     $criteria = $userFieldGateway->newQueryCriteria(true)
